Deep_Learning/Waste Classification/requirements.txt
streamlit
tensorflow
numpy
Pillow
pandas
opencv-python
scikit-learn
scipy
matplotlib